+++ changelog.md
@@ -1,4 +1,5 @@
 ## Bug fixes
 - toggling `enhancedTable` is possible through the `Legend API`
+- symbology toggles and layer visibility toggles sync up with `enhancedTable` rows and filter status
 
 ## New Features